llm_study-dev/web/requirements.txt
python-dotenv # 환경변수를 파일(.env)로 관리하는 라이브러리
streamlit # 웹 애플리케이션 개발 라이브러리
openai # 오픈AI API 클라이언트 호출 라이브러리
jupyter # 데이터 분석 및 시각화 툴
pandas # 데이터 조작 및 분석 라이브러리
groq # groq에서 지원하는 LLM을 API 클라이언트 호출 라이브러리
ollama
psycopg2-binary # postgresql에 접속할 수 있는 라이브러리 
sqlalchemy # streamlit과 postgresql를 연동하게 해주는 라이블러리
langchain # LLM을 이용한 애플리케이션 라이브러리 
langchain-community # 랭체인에서 사용한 올라마와 같은 서드파티 라이브러리 
langchain-core 
faiss-cpu
langchain-openai

